l8kyu
Module fundamentals

Modules categorized by Codewars labels - 8kyu *** Fundamentals ***

Functions

derive

This function takes two numbers as parameters, the first number being the coefficient, and the second number being the exponent. The exponent will never be 1, and neither number will ever be 0.

is_love

Timmy & Sarah think they are in love, but around where they live, they will only know once they pick a flower each. If one of the flowers has an even number of petals and the other has an odd number of petals it means they are in love. Write a function that will take the number of petals of each flower and return true if they are in love and false if they aren’t.

multi_table

Your goal is to return multiplication table for number that is always an integer from 1 to 10.

no_space

Write a function that removes the spaces from the string, then return the resultant string.

to_alternating_case

Define String.prototype.toAlternatingCase (or a similar function/method such as to_alternating_case/toAlternatingCase/ToAlternatingCase in your selected language; see the initial solution for details) such that each lowercase letter becomes uppercase and each uppercase letter becomes lowercase.